% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bhattacharyya.R
\name{BhattacharyyaDist}
\alias{BhattacharyyaCoef}
\alias{BhattacharyyaDist}
\title{Bhattacharyya Distance}
\usage{
BhattacharyyaDist(P, Q, dx)
}
\arguments{
\item{P}{y-values of a distribution}

\item{Q}{y-values of another distribution}

\item{dx}{delta x}
}
\description{
Calcuate Bhattacharyya Distance between two probability density functions.
}
\seealso{
\code{\link{KLDiv}} \code{\link{JSDiv}} \code{\link{JSDist}}
}

