% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entropy.R
\name{Entropy}
\alias{Entropy}
\title{Entropy}
\usage{
Entropy(P, dx)
}
\arguments{
\item{P}{y-values of a distribution}

\item{dx}{delta x}
}
\description{
calculate the entropy of a probability density function
}

