% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hellinger.R
\name{HellingerDist}
\alias{HellingerDist}
\title{Hellinger Distance}
\usage{
HellingerDist(P, Q, dx)
}
\arguments{
\item{P}{y-values of a distribution}

\item{Q}{y-values of another distribution}

\item{dx}{delta x}
}
\description{
Calcuate Hellinger Distance between two probability density functions.
}
\seealso{
\code{\link{BhattacharyyaDist}} \code{\link{KLDiv}} \code{\link{JSDiv}} \code{\link{JSDist}}
}

