% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density.R
\name{IntSig}
\alias{IntSig}
\title{IntSig}
\usage{
IntSig(intervals)
}
\arguments{
\item{intervals}{a vector of inter-event times in seconds}
}
\description{
Kernel Density Estimation for Log of Inter-event Times, This function receives
}
\examples{
intervals = c(1,5,2,3,15,2,354374,124135,352,234,14,5331)
KDE.int(intervals)

}

