% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSD.R
\name{JSDiv}
\alias{JSD}
\alias{JSDist}
\alias{JSDiv}
\title{Jensen-Shannon Divergence/Distance}
\usage{
JSDiv(P, Q, dx)
}
\arguments{
\item{P}{y-values of a distribution}

\item{Q}{y-values of another distribution}

\item{dx}{delta x}
}
\description{
Calcuate Jensen-Shannon divergence/distance
}
\seealso{
\code{\link{KLDiv}} \code{\link{BhattacharyyaDist}}
}

