% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KLD.R
\name{KLDiv}
\alias{KLD}
\alias{KLDiv}
\title{Kullback-Leibler divergence}
\usage{
KLDiv(P, Q, dx)
}
\arguments{
\item{P}{y-values of a distribution}

\item{Q}{y-values of another distribution}

\item{dx}{delta x}
}
\description{
calculate K-L divergence w/ trapezodial approximation
}

