% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Density.R, R/plot.R
\name{PlotEvent}
\alias{PlotEvent}
\title{Timestamp plotting}
\usage{
PlotEvent(timestamps, alpha = 1/4, ...)

PlotEvent(timestamps, alpha = 1/4, ...)
}
\arguments{
\item{UNIX}{timestamps/epochs of an event series}

\item{UNIX}{timestamps/epochs of an event series}
}
\description{
Plot timestamps with vertical lines

Plot timestamps with vertical lines
}

