% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{StripAndSplit}
\alias{StripAndSplit}
\title{Strip and Split}
\usage{
StripAndSplit(l, n.split = 10, n.strip = 0)
}
\arguments{
\item{l}{a vector}

\item{n.split}{the number of bins to split into}

\item{n.strip}{the number (or proportion) of elements to strip at each end}
}
\description{
strip the first and last elements, then split the rest
}

